Final processing of functionalities
