+++ package.json
@@ -42,7 +42,7 @@
 		"react-is": "^17.0.1",
 		"react-router-dom": "^5.2.0",
 		"styled-components": "^5.2.1",
-		"uuidv4": "^6.2.6",
+		"uuid": "^8.3.2",
 		"web-vitals": "^0.2.4",
 		"with-alpha-hex": "^1.0.6"
 	},
@@ -58,6 +58,7 @@
 		"@types/react-dom": "^16.9.8",
 		"@types/react-router-dom": "^5.1.7",
 		"@types/styled-components": "^5.1.7",
+		"@types/uuid": "^8.3.0",
 		"@typescript-eslint/eslint-plugin": "^4.0.0",
 		"@typescript-eslint/parser": "^4.0.0",
 		"babel-eslint": "^10.0.0",
