+++ src/components/Button/index.tsx
@@ -15,6 +15,8 @@ interface Props {
 	subtitle?: string
 	onClick?: (event: React.MouseEvent<HTMLButtonElement, MouseEvent>) => void
 	className?: string
+	name?: string
+	type?: 'button' | 'submit' | 'reset'
 }
 
 const Button: React.FC<Props> = ({
@@ -24,8 +26,15 @@ const Button: React.FC<Props> = ({
 	subtitle,
 	onClick,
 	className,
+	name,
+	type = 'button',
 }) => (
-	<StyledButton className={className} onClick={onClick}>
+	<StyledButton
+		className={className}
+		onClick={onClick}
+		name={name}
+		type={type}
+	>
 		{icon && <WrapperIcon>{React.createElement(icon)}</WrapperIcon>}
 		{title && <ButtonTitle>{title}</ButtonTitle>}
 		{titleSmall && <ButtonTitleSmall>{titleSmall}</ButtonTitleSmall>}
