--- src/components/ContentProfile/Email/styled.ts
@@ -1,7 +1,0 @@
-import styled from 'styled-components'
-
-export const EmailAccount = styled.span`
-display: block;
-font-size: 1.4em;
-text-align: center;
-`
