--- src/components/ContentProfile/SocialNetwork/Network/styled.ts
@@ -1,5 +1,0 @@
-import styled from 'styled-components'
-
-export const ListItem = styled.li``
-
-export const NetworkText = styled.span``
