--- src/components/ContentProfile/SocialNetwork/NetworkList/styled.ts
@@ -1,8 +1,0 @@
-import styled from 'styled-components'
-
-export const List = styled.ul`
-list-style: none;
-padding: 0;
-`
-
-export const NoNetworks = styled.div``
