+++ src/components/Profile/SocialNetwork/index.tsx
@@ -1,0 +1,12 @@
+import React from 'react'
+import { Title } from '../styled'
+import { NetworkForm } from './NetworkForm'
+import { NetworkList } from './NetworkList'
+
+export const SocialNetwork: React.FC = () => (
+	<>
+		<Title>Kde mě najdeš</Title>
+		<NetworkForm />
+		<NetworkList />
+	</>
+)
